renaming authicatedUser to UserPrincipal
